requirements.txt
# =========================
# Core Backend
# =========================
fastapi==0.110.0
uvicorn==0.29.0
httpx[http2]==0.25.2

# =========================
# Authentication & Security
# =========================
python-jose==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1

# =========================
# Database
# =========================
pymongo==4.6.2

# =========================
# Configuration
# =========================
pydantic==2.6.1
pydantic-settings==2.2.1

# =========================
# AI / LLM
# =========================
google-generativeai==0.3.2

# =========================
# Experiment Tracking
# =========================
mlflow==2.11.3
rich==13.7.0

# =========================
# Voice Processing
# =========================
openai-whisper==20231117
edge-tts==6.1.10

# =========================
# Vision / CNN
# =========================
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
pytesseract==0.3.10

# =========================
# Cloud / Storage
# =========================
boto3==1.34.34

# =========================
# Testing
# =========================
pytest==8.0.2
pytest-mock==3.12.0

# =========================
# HTTP / Utils
# =========================
requests==2.31.0
python-multipart==0.0.9
# =========================
# Validation
# =========================
email-validator==2.1.1
# =========================
# Async Testing
# =========================
pytest-asyncio==0.23.5
pytest-cov==4.1.0
anyio==4.2.0

black==24.10.0
